added HBase timestamp support https://github.com/hortonworks-spark/shc/issues/210

+++ core/src/test/scala/org/apache/spark/sql/SHC.scala
@@ -54,25 +54,30 @@ class SHC  extends FunSuite with BeforeAndAfterEach with BeforeAndAfterAll  with
   // private[spark] var columnFamilyStr = Bytes.toString(columnFamily)
 
   def defineCatalog(tName: String) = s"""{
-                                         |"table":{"namespace":"default", "name":"$tName"},
-                                         |"rowkey":"key",
-                                         |"columns":{
-                                              |"col0":{"cf":"rowkey", "col":"key", "type":"string"},
-                                              |"col1":{"cf":"cf1", "col":"col1", "type":"boolean"},
-                                              |"col2":{"cf":"cf2", "col":"col2", "type":"double"},
-                                              |"col3":{"cf":"cf3", "col":"col3", "type":"float"},
-                                              |"col4":{"cf":"cf4", "col":"col4", "type":"int"},
-                                              |"col5":{"cf":"cf5", "col":"col5", "type":"bigint"},
-                                              |"col6":{"cf":"cf6", "col":"col6", "type":"smallint"},
-                                              |"col7":{"cf":"cf7", "col":"col7", "type":"string"},
-                                              |"col8":{"cf":"cf8", "col":"col8", "type":"tinyint"}
-                                            |}
-                                         |}""".stripMargin
+                                        |"table":{"namespace":"default", "name":"$tName"},
+                                        |"rowkey":"key",
+                                        |"timestamp":"ts",
+                                        |"columns":{
+                                        |"col0":{"cf":"rowkey", "col":"key", "type":"string"},
+                                        |"colT":{"cf":"timestamp", "col":"ts", "type":"long"},
+                                        |"col1":{"cf":"cf1", "col":"col1", "type":"boolean"},
+                                        |"col2":{"cf":"cf2", "col":"col2", "type":"double"},
+                                        |"col3":{"cf":"cf3", "col":"col3", "type":"float"},
+                                        |"col4":{"cf":"cf4", "col":"col4", "type":"int"},
+                                        |"col5":{"cf":"cf5", "col":"col5", "type":"bigint"},
+                                        |"col6":{"cf":"cf6", "col":"col6", "type":"smallint"},
+                                        |"col7":{"cf":"cf7", "col":"col7", "type":"string"},
+                                        |"col8":{"cf":"cf8", "col":"col8", "type":"tinyint"}
+                                        |}
+                                        |}""".stripMargin
 
   @deprecated(since = "04.12.2017(dd/mm/year)", message = "use `defineCatalog` instead")
   def catalog = defineCatalog(tableName)
 
   override def beforeAll() {
+    if (sc != null) {
+      return
+    }
     val tempDir: File = Files.createTempDir
     tempDir.deleteOnExit
     htu.startMiniCluster
